requirements.txt
flask
python-dotenv
openai
gtts
langchain
langchain-openai